refactor: restructure layout-related components
