+++ src/components/Layout/Navbar/Navbar.module.css
@@ -1,0 +1,60 @@
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 16px;
+  background-color: var(--bg-white);
+  box-shadow: var(--box-shadow);
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 60px;
+  color: var(--primary-white);
+  z-index: 999;
+}
+
+.navbar__left {
+  display: flex;
+}
+
+.navbar__text {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: var(--primary-black);
+  font-size: var(--font-size);
+  font-weight: var(--font-bold-600);
+}
+
+.navbar__left__logo {
+  display: flex;
+  align-items: center;
+  width: 40px;
+  height: 40px;
+  margin-right: 10px;
+}
+
+.navbar__left__logo img {
+  width: 35px;
+  height: 35px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.navbar__right {
+  padding: 10px 0px;
+}
+
+.navbar__right img {
+  display: flex;
+  align-items: center;
+  height: 40px;
+  width: 40px;
+  object-fit: cover;
+  border: 1px solid var(--primary-red);
+  border-radius: 10px;
+  cursor: pointer;
+}
