+++ src/components/Layout/Sidebar/Sidebar.module.css
@@ -1,4 +1,3 @@
-/* Sidebar.module.css */
 .sidebar {
   position: fixed;
   left: 0;
@@ -35,8 +34,8 @@
   border: 1px solid var(--primary-red);
 }
 
-.sidebar__container{
-  margin-left:10px;
+.sidebar__container {
+  margin-left: 10px;
 }
 
 .sidebar__container p {
@@ -44,7 +43,7 @@
   color: var(--primary-black-3);
 }
 
-.text_bold_500{
+.text_bold_500 {
   font-weight: var(--font-bold);
 }
 
@@ -78,8 +77,8 @@
   background-color: var(--primary-red);
 }
 
-.sidebar__text{
-  margin-left:20px;
+.sidebar__text {
+  margin-left: 20px;
 }
 
 .sidebar__menu__active,
@@ -98,7 +97,7 @@
   align-items: center;
   background-color: var(--bg-white);
   position: absolute;
-  left: 5; 
+  left: 5;
 }
 
 .sidebar__menu__active .sidebar__icon,
