+++ src/components/Layout/index.js
@@ -1,7 +1,6 @@
-// Layout.js
 import React from "react";
-import Navbar from "../Navbar";
-import Sidebar from "../Sidebar";
+import Navbar from "./Navbar";
+import Sidebar from "./Sidebar";
 
 const Layout = ({ children }) => {
   return (
