+++ src/feature/Dashboard/index.jsx
@@ -1,4 +1,3 @@
-// Dashboard.js
 import React from "react";
 import Layout from "../../components/Layout";
 import EmployeesCard from "../../components/cards/EmployeeCard";
